fix: possibility of long-staled data
This is interesting because `getStaticProps` function will be called
once in the build time, which means if there is no new deployment for
a week and during that week some change happened to the data, it
wont be shown in the website until next build happens. Which is
unfortunate because you never know when people get some free time to
make some updates on the website or maybe there is no need to change.
Usually the data fetching strategy must consider the duration for
which the data fresh/valid for the user to see.

One Solution could be use `getServerSideProps` but it may make the
TTFB long and ruin the UX. Another option is to use `revalidate`
option which is part of the return and is `false` by default, can
be set to number of seconds to re-validate (hence rebuild) the page
every time the number of seconds passed after last build which should
make sure the data is semi-fresh when a user visits the page.

See [Next.js docs](https://nextjs.org/docs/basic-features/data-fetching#getstaticprops-static-generation)